+++ tests/models/language/pooling_mteb_test/test_gte.py
@@ -5,32 +5,36 @@ import pytest
 
 from tests.models.language.pooling.embed_utils import correctness_test_embed_models
 from tests.models.utils import (
+    CLSPoolingEmbedModelInfo,
+    CLSPoolingRerankModelInfo,
     EmbedModelInfo,
+    LASTPoolingEmbedModelInfo,
     RerankModelInfo,
 )
 
-from .mteb_embed_utils import mteb_test_embed_models
-from .mteb_score_utils import mteb_test_rerank_models
+from .mteb_utils import mteb_test_embed_models, mteb_test_rerank_models
 
 MODELS = [
     ########## BertModel
-    EmbedModelInfo(
+    CLSPoolingEmbedModelInfo(
         "thenlper/gte-large",
         mteb_score=0.76807651,
         architecture="BertModel",
-        pooling_type="MEAN",
-        attn_type="encoder_only",
-        is_prefix_caching_supported=False,
-        is_chunked_prefill_supported=False,
         enable_test=True,
     ),
-    EmbedModelInfo("thenlper/gte-base", architecture="BertModel", enable_test=False),
-    EmbedModelInfo("thenlper/gte-small", architecture="BertModel", enable_test=False),
-    EmbedModelInfo(
+    CLSPoolingEmbedModelInfo(
+        "thenlper/gte-base", architecture="BertModel", enable_test=False
+    ),
+    CLSPoolingEmbedModelInfo(
+        "thenlper/gte-small", architecture="BertModel", enable_test=False
+    ),
+    CLSPoolingEmbedModelInfo(
         "thenlper/gte-large-zh", architecture="BertModel", enable_test=False
     ),
-    EmbedModelInfo("thenlper/gte-base-zh", architecture="BertModel", enable_test=False),
-    EmbedModelInfo(
+    CLSPoolingEmbedModelInfo(
+        "thenlper/gte-base-zh", architecture="BertModel", enable_test=False
+    ),
+    CLSPoolingEmbedModelInfo(
         "thenlper/gte-small-zh", architecture="BertModel", enable_test=False
     ),
     ########### NewModel
@@ -39,64 +43,48 @@ MODELS = [
     # - whether to use token_type_embeddings
     # - whether to use context expansion
     # So only test one (the most widely used) model
-    EmbedModelInfo(
+    CLSPoolingEmbedModelInfo(
         "Alibaba-NLP/gte-multilingual-base",
         architecture="GteNewModel",
         mteb_score=0.775074696,
         hf_overrides={"architectures": ["GteNewModel"]},
-        pooling_type="CLS",
-        attn_type="encoder_only",
-        is_prefix_caching_supported=False,
-        is_chunked_prefill_supported=False,
         enable_test=True,
     ),
-    EmbedModelInfo(
+    CLSPoolingEmbedModelInfo(
         "Alibaba-NLP/gte-base-en-v1.5",
         architecture="GteNewModel",
         hf_overrides={"architectures": ["GteNewModel"]},
         enable_test=False,
     ),
-    EmbedModelInfo(
+    CLSPoolingEmbedModelInfo(
         "Alibaba-NLP/gte-large-en-v1.5",
         architecture="GteNewModel",
         hf_overrides={"architectures": ["GteNewModel"]},
         enable_test=False,
     ),
     ########### Qwen2ForCausalLM
-    EmbedModelInfo(
+    LASTPoolingEmbedModelInfo(
         "Alibaba-NLP/gte-Qwen2-1.5B-instruct",
         mteb_score=0.758473459018872,
         architecture="Qwen2ForCausalLM",
-        pooling_type="LAST",
-        attn_type="encoder_only",
-        is_prefix_caching_supported=False,
-        is_chunked_prefill_supported=False,
         enable_test=True,
     ),
     ########## ModernBertModel
-    EmbedModelInfo(
+    CLSPoolingEmbedModelInfo(
         "Alibaba-NLP/gte-modernbert-base",
         mteb_score=0.748193353,
         architecture="ModernBertModel",
-        pooling_type="CLS",
-        attn_type="encoder_only",
-        is_prefix_caching_supported=False,
-        is_chunked_prefill_supported=False,
         enable_test=True,
     ),
     ########## Qwen3ForCausalLM
-    EmbedModelInfo(
+    LASTPoolingEmbedModelInfo(
         "Qwen/Qwen3-Embedding-0.6B",
         mteb_score=0.771163695,
         architecture="Qwen3ForCausalLM",
         dtype="float32",
-        pooling_type="LAST",
-        attn_type="decoder",
-        is_prefix_caching_supported=True,
-        is_chunked_prefill_supported=True,
         enable_test=True,
     ),
-    EmbedModelInfo(
+    LASTPoolingEmbedModelInfo(
         "Qwen/Qwen3-Embedding-4B",
         architecture="Qwen3ForCausalLM",
         dtype="float32",
@@ -105,26 +93,18 @@ MODELS = [
 ]
 
 RERANK_MODELS = [
-    RerankModelInfo(
+    CLSPoolingRerankModelInfo(
         # classifier_pooling: mean
         "Alibaba-NLP/gte-reranker-modernbert-base",
         mteb_score=0.33386,
         architecture="ModernBertForSequenceClassification",
-        pooling_type="CLS",
-        attn_type="encoder_only",
-        is_prefix_caching_supported=False,
-        is_chunked_prefill_supported=False,
         enable_test=True,
     ),
-    RerankModelInfo(
+    CLSPoolingRerankModelInfo(
         "Alibaba-NLP/gte-multilingual-reranker-base",
         mteb_score=0.33062,
         architecture="GteNewForSequenceClassification",
         hf_overrides={"architectures": ["GteNewForSequenceClassification"]},
-        pooling_type="CLS",
-        attn_type="encoder_only",
-        is_prefix_caching_supported=False,
-        is_chunked_prefill_supported=False,
         enable_test=True,
     ),
 ]
