+++ tests/models/language/pooling_mteb_test/test_intfloat.py
@@ -3,44 +3,40 @@
 import pytest
 
 from tests.models.language.pooling.embed_utils import correctness_test_embed_models
-from tests.models.utils import EmbedModelInfo
+from tests.models.utils import CLSPoolingEmbedModelInfo, EmbedModelInfo
 
-from .mteb_embed_utils import mteb_test_embed_models
+from .mteb_utils import mteb_test_embed_models
 
 MODELS = [
     ########## BertModel
-    EmbedModelInfo(
+    CLSPoolingEmbedModelInfo(
         "intfloat/e5-small",
         architecture="BertModel",
         mteb_score=0.742285423,
-        pooling_type="MEAN",
-        attn_type="encoder_only",
-        is_prefix_caching_supported=False,
-        is_chunked_prefill_supported=False,
         enable_test=True,
     ),
-    EmbedModelInfo("intfloat/e5-base", architecture="BertModel", enable_test=False),
-    EmbedModelInfo("intfloat/e5-large", architecture="BertModel", enable_test=False),
-    EmbedModelInfo(
+    CLSPoolingEmbedModelInfo(
+        "intfloat/e5-base", architecture="BertModel", enable_test=False
+    ),
+    CLSPoolingEmbedModelInfo(
+        "intfloat/e5-large", architecture="BertModel", enable_test=False
+    ),
+    CLSPoolingEmbedModelInfo(
         "intfloat/multilingual-e5-small", architecture="BertModel", enable_test=False
     ),
     ########## XLMRobertaModel
-    EmbedModelInfo(
+    CLSPoolingEmbedModelInfo(
         "intfloat/multilingual-e5-base",
         architecture="XLMRobertaModel",
         mteb_score=0.779325955,
-        pooling_type="MEAN",
-        attn_type="encoder_only",
-        is_prefix_caching_supported=False,
-        is_chunked_prefill_supported=False,
         enable_test=True,
     ),
-    EmbedModelInfo(
+    CLSPoolingEmbedModelInfo(
         "intfloat/multilingual-e5-large",
         architecture="XLMRobertaModel",
         enable_test=False,
     ),
-    EmbedModelInfo(
+    CLSPoolingEmbedModelInfo(
         "intfloat/multilingual-e5-large-instruct",
         architecture="XLMRobertaModel",
         enable_test=False,
