+++ tests/models/language/pooling_mteb_test/test_nemotron.py
@@ -3,39 +3,29 @@
 
 import pytest
 
-from tests.models.language.pooling_mteb_test.mteb_embed_utils import (
-    mteb_test_embed_models,
-)
-from tests.models.language.pooling_mteb_test.mteb_score_utils import (
-    mteb_test_rerank_models,
-)
 from tests.models.utils import (
     EmbedModelInfo,
+    LASTPoolingEmbedModelInfo,
+    LASTPoolingRerankModelInfo,
     RerankModelInfo,
 )
 
+from .mteb_utils import mteb_test_embed_models, mteb_test_rerank_models
+
 EMBEDDING_MODELS = [
-    EmbedModelInfo(
+    LASTPoolingEmbedModelInfo(
         "nvidia/llama-nemotron-embed-1b-v2",
         architecture="LlamaBidirectionalModel",
         mteb_score=0.689164662128673,
-        pooling_type="MEAN",
-        attn_type="encoder_only",
-        is_prefix_caching_supported=False,
-        is_chunked_prefill_supported=False,
     )
 ]
 
 RERANK_MODELS = [
-    RerankModelInfo(
+    LASTPoolingRerankModelInfo(
         "nvidia/llama-nemotron-rerank-1b-v2",
         architecture="LlamaBidirectionalForSequenceClassification",
         chat_template_name="nemotron-rerank.jinja",
         mteb_score=0.33994,
-        pooling_type="MEAN",
-        attn_type="encoder_only",
-        is_prefix_caching_supported=False,
-        is_chunked_prefill_supported=False,
     ),
 ]
