+++ tests/models/language/pooling_mteb_test/test_snowflake_arctic_embed.py
@@ -4,82 +4,62 @@
 import pytest
 
 from tests.models.language.pooling.embed_utils import correctness_test_embed_models
-from tests.models.utils import EmbedModelInfo
+from tests.models.utils import CLSPoolingEmbedModelInfo, EmbedModelInfo
 
-from .mteb_embed_utils import mteb_test_embed_models
+from .mteb_utils import mteb_test_embed_models
 
 MODELS = [
-    EmbedModelInfo(
+    CLSPoolingEmbedModelInfo(
         "Snowflake/snowflake-arctic-embed-xs",
         is_matryoshka=False,
         architecture="BertModel",
         mteb_score=0.714927797,
-        pooling_type="CLS",
-        attn_type="encoder_only",
-        is_prefix_caching_supported=False,
-        is_chunked_prefill_supported=False,
         enable_test=True,
     ),
-    EmbedModelInfo(
+    CLSPoolingEmbedModelInfo(
         "Snowflake/snowflake-arctic-embed-s",
         is_matryoshka=False,
         architecture="BertModel",
         enable_test=False,
     ),
-    EmbedModelInfo(
+    CLSPoolingEmbedModelInfo(
         "Snowflake/snowflake-arctic-embed-m",
         is_matryoshka=False,
         architecture="BertModel",
         enable_test=False,
     ),
-    EmbedModelInfo(
+    CLSPoolingEmbedModelInfo(
         "Snowflake/snowflake-arctic-embed-m-long",
         is_matryoshka=False,
         architecture="NomicBertModel",
         mteb_score=0.681146831,
-        pooling_type="CLS",
-        attn_type="encoder_only",
-        is_prefix_caching_supported=False,
-        is_chunked_prefill_supported=False,
         enable_test=True,
     ),
-    EmbedModelInfo(
+    CLSPoolingEmbedModelInfo(
         "Snowflake/snowflake-arctic-embed-l",
         is_matryoshka=False,
         architecture="BertModel",
         enable_test=False,
     ),
-    EmbedModelInfo(
+    CLSPoolingEmbedModelInfo(
         "Snowflake/snowflake-arctic-embed-m-v1.5",
         is_matryoshka=True,
         architecture="BertModel",
         mteb_score=0.649088363,
-        pooling_type="CLS",
-        attn_type="encoder_only",
-        is_prefix_caching_supported=False,
-        is_chunked_prefill_supported=False,
         enable_test=True,
     ),
-    EmbedModelInfo(
+    CLSPoolingEmbedModelInfo(
         "Snowflake/snowflake-arctic-embed-l-v2.0",
         is_matryoshka=True,
         architecture="XLMRobertaModel",
         mteb_score=0.712258299,
-        pooling_type="CLS",
-        attn_type="encoder_only",
-        is_prefix_caching_supported=False,
-        is_chunked_prefill_supported=False,
         enable_test=True,
     ),
-    EmbedModelInfo(
+    CLSPoolingEmbedModelInfo(
         "Snowflake/snowflake-arctic-embed-m-v2.0",
         is_matryoshka=True,
         architecture="GteModel",
         mteb_score=0.706622444,
-        pooling_type="CLS",
-        attn_type="encoder_only",
-        is_prefix_caching_supported=False,
-        is_chunked_prefill_supported=False,
         enable_test=True,
     ),
 ]
