+++ tests/models/language/pooling_mteb_test/test_st_projector.py
@@ -3,31 +3,25 @@
 import pytest
 
 from tests.models.utils import (
+    CLSPoolingEmbedModelInfo,
     EmbedModelInfo,
+    LASTPoolingEmbedModelInfo,
 )
 
-from .mteb_embed_utils import mteb_test_embed_models
+from .mteb_utils import mteb_test_embed_models
 
 # ST models with projector (Dense) layers
 ST_PROJECTOR_MODELS = [
-    EmbedModelInfo(
+    CLSPoolingEmbedModelInfo(
         "TencentBAC/Conan-embedding-v1",
         architecture="BertModel",
         mteb_score=0.688611955,
-        pooling_type="MEAN",
-        attn_type="encoder_only",
-        is_prefix_caching_supported=False,
-        is_chunked_prefill_supported=False,
         enable_test=True,
     ),
-    EmbedModelInfo(
+    LASTPoolingEmbedModelInfo(
         "google/embeddinggemma-300m",
         architecture="Gemma3TextModel",
         mteb_score=0.7473819294684156,
-        pooling_type="MEAN",
-        attn_type="encoder_only",
-        is_prefix_caching_supported=False,
-        is_chunked_prefill_supported=False,
         enable_test=True,
         dtype="float32",
     ),
